+++ R/fit_helpers.R
@@ -128,6 +128,7 @@ xy_xy <- function(
       y_name <- colnames(env$y)
     }
   }
+
   res$preproc <- list(y_var = y_name, x_names = colnames(env$x))
   res$elapsed <- list(elapsed = elapsed, print = control$verbosity > 1L)
   res
@@ -217,9 +218,9 @@ xy_form <- function(object, env, control, ...) {
   } else {
     if (is.atomic(env$y)) {
       data_obj$y_var <- character(0)
+    } else {
+      data_obj$y_var <- colnames(env$y)
     }
-
-    data_obj$y_var <- colnames(env$y)
   }
 
   res$preproc <- data_obj[c("x_var", "y_var")]